+++ CHANGELOG.md
@@ -74,6 +74,7 @@
     - Fixed not working mappings caused by use of numpad keys (#220)
     - Added view-commands using smooth transitions (#139)
     - Added "Add module" command
+    - Added scroll-commands (#252)
     - Added tooltips for mapped commands
 - Module [TRANSIT](./docs/Transit.md)
     - Added "Phase"-mode for CV-input which scans continously through snapshots (#182)